refactor: SimpleModal 리팩토링, Modal 컴포넌트 개발
