+++ src/components/common/modal/SimpleModal.tsx
@@ -1,19 +1,8 @@
 import React from 'react';
 import './index.scss';
-import { BottomSheet } from 'react-spring-bottom-sheet';
-import 'reactjs-popup/dist/index.css';
-import { Container, Dialog, Typography } from '@material-ui/core';
-import useWindowDimensions from '@src/hooks/useWindowDimensions';
-
-interface ISimpleModalProps {
-	open: boolean;
-	onClose: (param: boolean) => void;
-	children: JSX.Element;
-	footer?: boolean;
-	title?: string;
-	titleStyle?: React.CSSProperties;
-	height?: string;
-}
+import { Container, Typography } from '@material-ui/core';
+import { ISimpleModalProps } from './types';
+import Modal from './Modal';
 
 const SimpleModal = ({
 	open,
@@ -24,16 +13,9 @@ const SimpleModal = ({
 	titleStyle,
 	height,
 }: ISimpleModalProps): JSX.Element => {
-	const { width } = useWindowDimensions();
-
-	const isDesktop = width > 1024;
-
-	const Content = () => {
-		return (
-			<Container
-				className="simple-modal-container"
-				style={{ width: isDesktop ? '20rem' : '100%', height: isDesktop ? 'auto' : height }}
-			>
+	return (
+		<Modal open={open} onClose={onClose} height={height}>
+			<>
 				{title && (
 					<Container>
 						<Typography className="simple-modal-title" style={titleStyle}>
@@ -49,18 +31,8 @@ const SimpleModal = ({
 						</Typography>
 					</Container>
 				)}
-			</Container>
-		);
-	};
-
-	return isDesktop ? (
-		<Dialog open={open}>
-			<Content />
-		</Dialog>
-	) : (
-		<BottomSheet open={open} onDismiss={() => onClose(false)} className="simple-modal-bottom-sheet">
-			<Content />
-		</BottomSheet>
+			</>
+		</Modal>
 	);
 };
