+++ src/components/common/modal/index.scss
@@ -1,4 +1,8 @@
-.simple-modal-bottom-sheet {
+.modal-container {
+    padding: 0 !important;
+}
+
+.modal-bottom-sheet {
     > div {
         border-top-left-radius: .75rem;
         border-top-right-radius: .75rem;
@@ -6,19 +10,11 @@
     }
 }
 
-.simple-modal-container {
-    padding: 0 !important;
-}
-
 .simple-modal-title {
     padding-top: 2rem;
     font-size: 1.125rem !important;
     font-weight: bold !important;
     text-align: center;
-}
-
-.simple-modal-content {
-    
 }
 
 .simple-modal-footer {
